theoretical option started for ddd update method
i don't think this makes sense in practice though.in this case at least, it would be nice to pull in logic for a single 'update' action, but while successful on the update patient command, on the patch patient command, there is some work that has to be done in between some of this update function. there would be theoretical nice coupling to have here, but i think in practice it's more maintainable to have the info isolated to the feature. if i'ts domain logic like creating an internal id with a service though, that should probably go in here... let's try that now